updated to 5.4.100; the Maps App now shows the whole map on startup again, and the Windows demo compiles again after recent changes to the API

+++ src/demo/windows_demo/FindDialog.cpp
@@ -33,7 +33,7 @@ iFramework(aFramework)
 
 CartoType::CMapObjectArray CFindTextDialog::FoundObjectArray()
     {
-    if (iListBoxIndex >= 0 && iListBoxIndex <= iMapObjectGroupArray.size())
+    if (iListBoxIndex >= 0 && size_t(iListBoxIndex) <= iMapObjectGroupArray.size())
         return std::move(iMapObjectGroupArray[iListBoxIndex].iMapObjectArray);
 
     CartoType::CMapObjectArray a;
